Cache book covers to speed rescans

+++ lazylibrarian/bookcovers.py
@@ -28,9 +28,9 @@ def getBookCovers(bookids=None):
             title = formatter.safe_unicode(item['BookName']).encode('utf-8')
             author = formatter.safe_unicode(item['AuthorName']).encode('utf-8')
             safeparams = urllib.quote_plus("%s %s" % (author, title))
-            URL4="https://www.google.com/search?as_st=y&tbm=isch&as_q=" + safeparams + "+ebook&tbs=isz:l,ift:jpg&gws_rd=cr&ei=Ff30Vo_HOaWuygO13bvYBQ"
+            URL="https://www.google.com/search?as_st=y&tbm=isch&as_q=" + safeparams + "+ebook&tbs=isz:l,ift:jpg&gws_rd=cr&ei=Ff30Vo_HOaWuygO13bvYBQ"
     
-            request = urllib2.Request(URL4)
+            request = urllib2.Request(URL)
             #if lazylibrarian.PROXY_HOST:
             #    request.set_proxy(lazylibrarian.PROXY_HOST, lazylibrarian.PROXY_TYPE)
             # google insists on having a user-agent
@@ -44,39 +44,42 @@ def getBookCovers(bookids=None):
                 os.makedirs(cachedir)
             coverfile = os.path.join(cachedir, hashID + '.jpg')
             link = 'images' + os.sep + 'cache' + os.sep + hashID + '.jpg'
-
-            try:
-                resp = urllib2.urlopen(request, timeout=30)
-                if str(resp.getcode()).startswith("2"):
-                    # (200 OK etc)
-                    source_page = resp.read()
-                    #with open(coverfile + 'src', 'wb') as img:
-                    #    img.write(source_page)
-                    #print coverfile + 'src'
-                    try:
-                        img = source_page.split('url?q=')[1].split('">')[1].split('src="')[1].split('"')[0]
-                    except IndexError:
-                        img = "Not found in source page"
-                    if img.startswith('http'):
-                        request = urllib2.Request(img)
-                        request.add_header('User-Agent', USER_AGENT)
+            
+            if not os.path.isfile(coverfile):
+                # use cached image if possible to speed up 
+                # refreshactiveauthors and librarysync re-runs
+                try:
+                    resp = urllib2.urlopen(request, timeout=30)
+                    if str(resp.getcode()).startswith("2"):
+                        # (200 OK etc)
+                        source_page = resp.read()
+                        #with open(coverfile + 'src', 'wb') as img:
+                        #    img.write(source_page)
+                        #print coverfile + 'src'
                         try:
-                            resp = urllib2.urlopen(request, timeout=30)
-                            if str(resp.getcode()).startswith("2"):
-                                # (200 OK etc)
-                                with open(coverfile, 'wb') as img:
-                                    img.write(resp.read())
-                                update = True
-                            else:
-                                logger.debug("Error getting image: %s" % str(resp))
-                        except (urllib2.HTTPError, urllib2.URLError, socket.timeout) as e:
-                            logger.debug("Error getting image : %s" % e.reason)
-                    else:
-                        logger.debug("Error getting url for %s [%s]" % (safeparams, img))                      
-            except (urllib2.HTTPError, urllib2.URLError, socket.timeout) as e:
-                logger.debug("Error getting source page : %s" % e.reason)
+                            img = source_page.split('url?q=')[1].split('">')[1].split('src="')[1].split('"')[0]
+                        except IndexError:
+                            img = "image not found in source page"
+                        if img.startswith('http'):
+                            request = urllib2.Request(img)
+                            request.add_header('User-Agent', USER_AGENT)
+                            try:
+                                resp = urllib2.urlopen(request, timeout=30)
+                                if str(resp.getcode()).startswith("2"):
+                                    # (200 OK etc)
+                                    with open(coverfile, 'wb') as img:
+                                        img.write(resp.read())
+                                    update = True
+                                else:
+                                    logger.debug("Error getting image: %s" % str(resp))
+                            except (urllib2.HTTPError, urllib2.URLError, socket.timeout) as e:
+                                logger.debug("Error getting image : %s" % e.reason)
+                        else:
+                            logger.debug("Error getting url for %s [%s]" % (safeparams, img))                      
+                except (urllib2.HTTPError, urllib2.URLError, socket.timeout) as e:
+                    logger.debug("Error getting source page : %s" % e.reason)
 
-            # image downloaded, now update the link
+            # image downloaded, or was already there, now update the link
             logger.debug("Found GoogleImage cover for %s %s" % (author, title))
             myDB.action('update books set BookImg="%s" where BookID="%s"' % (link, bookid))
     if num > 1:
